+++ leilao/src/main/java/com/example/leilao/service/VeiculoService.java
@@ -1,5 +1,6 @@
 package com.example.leilao.service;
 
+import com.example.leilao.exception.ResourceNotFoundException;
 import com.example.leilao.model.Veiculo;
 import com.example.leilao.repository.VeiculoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,27 +15,25 @@ public class VeiculoService {
     @Autowired
     private VeiculoRepository veiculoRepository;
 
-    // Criar ou atualizar um veículo
     public Veiculo criarOuAtualizarVeiculo(Veiculo veiculo) {
         return veiculoRepository.save(veiculo);
     }
 
-    // Buscar veículo por ID
     public Optional<Veiculo> buscarVeiculoPorId(Long id) {
         return veiculoRepository.findById(id);
     }
 
-    // Listar todos os veículos
     public List<Veiculo> listarVeiculos() {
         return veiculoRepository.findAll();
     }
 
-    // Excluir veículo por ID
     public void excluirVeiculo(Long id) {
+        if (!veiculoRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Veículo com ID " + id + " não encontrado.");
+        }
         veiculoRepository.deleteById(id);
     }
-    
-    // Buscar veículos por tipo
+
     public List<Veiculo> buscarVeiculosPorTipo(String tipo) {
         return veiculoRepository.findByTipo(tipo);
     }
